Fixes link deletion.

+++ src/link/client/ReactItem.ts
@@ -12,6 +12,7 @@ class Component extends react.Component<any, any> {
     }
   }
   private handleDelete() {
+    if (!confirm('Are you sure you want to delete this link?')) return;
     this.props.onDelete(this.props.id);
   }
   render() {
@@ -25,11 +26,11 @@ class Component extends react.Component<any, any> {
         dom.div({}, [dom.a({className: 'short_url text-primary', href: short_url}, [short_url])]),
         dom.div({}, [dom.a({className: 'long_url small text-muted', href: this.props.long_url}, [this.props.long_url])])
       ]),
-      dom.span({className: 'col col-xs-1 link_action_col pull-right'}, [
+      this.props.owner ? dom.span({className: 'col col-xs-1 link_action_col pull-right'}, [
         dom.span({className: 'delete btn dtn-danger', onClick: this.handleDelete.bind(this)}, [
           dom.span({className: 'glyphicon glyphicon-remove'})
         ])
-      ])
+      ]) : null
     ]);
   }
 }
